Price adjustment/change: My campaigns conversion rate (Valuation to Appointment) has dropped down from 16% (Dec, 2017) to 5% (Mar 2018, till present)…
We have agreed on adding the following to Valuations in Admin:
We will add a column to show actual prices from the Valuation Engine regardless of what was shown to the customer.
We will also add a column to show the percentage and value affected by the Valuation Configurations.
With the changes above, we can easily see the cars that do not have prices because of the Valuation Engine versus those that have prices from the Valuation Engine but got affected by erroneous rules in the Valuation Configurations.